refactor(agent): compaction to its D6 home; builtin/ dissolved (#606)
Two moves the plan ordered and one rename it justified.

The mechanism moves from core/execution/compaction.ts to
src/compaction/compact.ts — D6 verbatim: compaction lives in
agent/src/compaction/ as pure modules, strategy opinions come in as
config. The namespace drops its InMemory prefix: there is no other
compactor, and the prefix implied a durable sibling that never existed.

The seam adapter (the policy registration wiring run.completion.pre +
run.replace_messages) moves beside the mechanism as compaction/policy.ts
rather than to the product: pinning WHERE compaction applies is D8
mechanism — "application stays at one deterministic seam" — and only
the strategy (thresholds, onSummarize) is opinion, which still arrives
as config. That empties core/policy/builtin/, so the directory is gone;
Phase 4 rule 3 is resolved by dissolution rather than by a guard.

Fallout the gates caught while landing it: the check-deps depth rule
rejected the transient deep import, which is what forced the policy
file's real home; the dead-export ratchet trimmed the new barrel twice
(CompactionBoundaryError, then the Compaction namespace) down to the
surface something actually consumes; and the root InMemoryCompactor
re-export had zero external importers, so it is deleted, not renamed.

Also removes a lint-guards canonicalPolicyRequiredFiles entry naming a
file that does not exist (builtin/messenger-allow-pattern.ts) — an
entry for a nonexistent file is never scanned and enforces nothing.

defaultRegistry survives this PR; it dies when the registration moves
to openomni like idle-nudge, budget-nudge, and tool-permission did.
