+++ packages/agent/bench/index.ts
@@ -2,7 +2,7 @@
 import { mkdirSync, writeFileSync } from "node:fs";
 import { Bench } from "tinybench";
 import type { Message } from "@openomni/protocol";
-import { InMemoryCompactor } from "../src/core/execution/compaction.ts";
+import { Compaction } from "../src/compaction/compact.ts";
 import { noopSink } from "@openomni/telemetry";
 
 /** The bench stands in for one run. It measures compaction, not reporting. */
@@ -32,7 +32,7 @@ for (const size of [20, 100, 500]) {
   bench.add(
     `compaction/${size}-messages`,
     async () => {
-      await InMemoryCompactor.compact(messages, compactionOptions, BENCH_TRACE, BENCH_EVENTS);
+      await Compaction.compact(messages, compactionOptions, BENCH_TRACE, BENCH_EVENTS);
     },
     { async: true },
   );
@@ -41,7 +41,7 @@ for (const size of [20, 100, 500]) {
 bench.add(
   "compaction/should-compact",
   () => {
-    InMemoryCompactor.shouldCompact(104000, compactionOptions);
+    Compaction.shouldCompact(104000, compactionOptions);
   },
   { async: false },
 );
