+++ packages/agent/src/compaction/index.ts
@@ -1,0 +1,2 @@
+export type { CompactionOptions } from "./compact";
+export { createCompactionPolicy } from "./policy";
