--- packages/agent/src/core/policy/builtin/index.ts
@@ -1,1 +1,0 @@
-export { createCompactionPolicy } from "./compaction";
